+++ include/standardese/md_custom.hpp
@@ -32,6 +32,33 @@ namespace standardese
         friend detail::md_ptr_access;
     };
 
+    class md_code_block;
+
+    class md_code_block_advanced final : public md_leave
+    {
+    public:
+        static md_entity::type get_entity_type() STANDARDESE_NOEXCEPT
+        {
+            return md_entity::code_block_advanced_t;
+        }
+
+        static md_ptr<md_code_block_advanced> make(const md_code_block& cb);
+        static md_ptr<md_code_block_advanced> make(const md_entity& parent, const char* code,
+                                                   const char* lang);
+
+    protected:
+        md_entity_ptr do_clone(const md_entity* parent) const override;
+
+    private:
+        md_code_block_advanced(const md_entity& parent, const char* code, const char* lang);
+
+        md_code_block_advanced(const md_entity& parent, cmark_node* node);
+
+        friend detail::md_ptr_access;
+    };
+
+    class md_list_item;
+
     class md_inline_documentation final : public md_container
     {
     public:
@@ -43,9 +70,13 @@ namespace standardese
         static md_ptr<md_inline_documentation> make(const md_entity&   parent,
                                                     const std::string& heading);
 
+        md_entity& add_entity(md_entity_ptr entity) override;
+
         // adds the contents of all paragraphs in container
         // returns false if any child was not a paragraph
         bool add_item(const char* name, const char* id, const md_container& container);
+
+        md_list_item& get_item() STANDARDESE_NOEXCEPT;
 
         bool empty() const STANDARDESE_NOEXCEPT;