+++ include/standardese/output.hpp
@@ -9,7 +9,10 @@
 #include <string>
 #include <ostream>
 
+#include <spdlog/fmt/fmt.h>
+
 #include <standardese/md_blocks.hpp>
+#include <standardese/md_custom.hpp>
 #include <standardese/noexcept.hpp>
 #include <standardese/output_format.hpp>
 #include <standardese/output_stream.hpp>
@@ -41,13 +44,18 @@ namespace standardese
     class code_block_writer
     {
     public:
-        code_block_writer(const md_entity& parent) STANDARDESE_NOEXCEPT : parent_(&parent)
+        code_block_writer(const md_entity& parent, bool use_advanced) STANDARDESE_NOEXCEPT
+            : parent_(&parent),
+              use_advanced_(use_advanced)
         {
         }
 
-        md_ptr<md_code_block> get_code_block(const char* fence = "cpp")
+        md_entity_ptr get_code_block(const char* fence = "cpp")
         {
-            return md_code_block::make(*parent_, stream_.get_string().c_str(), fence);
+            if (use_advanced_)
+                return md_code_block_advanced::make(*parent_, stream_.get_string().c_str(), fence);
+            else
+                return md_code_block::make(*parent_, stream_.get_string().c_str(), fence);
         }
 
         void indent(unsigned width)
@@ -65,34 +73,46 @@ namespace standardese
             stream_.remove_trailing_line();
         }
 
-        code_block_writer& fill(std::size_t size, char c)
+        code_block_writer& fill_ws(std::size_t size)
         {
             for (std::size_t i = 0u; i != size; ++i)
-                stream_.write_char(c);
+                stream_.write_char(' ');
             return *this;
         }
 
         code_block_writer& operator<<(const char* str)
         {
-            stream_.write_str(str, std::strlen(str));
+            write_str_escaped(str, std::strlen(str));
             return *this;
         }
 
         code_block_writer& operator<<(const std::string& str)
         {
-            stream_.write_str(str.c_str(), str.size());
+            write_str_escaped(str.c_str(), str.size());
             return *this;
         }
 
         code_block_writer& operator<<(const string& str)
         {
-            stream_.write_str(str.c_str(), str.length());
+            write_str_escaped(str.c_str(), str.length());
+            return *this;
+        }
+
+        code_block_writer& write_link(bool top_level, const string& str, const string& unique_name)
+        {
+            if (str.empty())
+                return *this;
+            else if (!use_advanced_ || top_level || unique_name.empty())
+                return *this << str;
+            write_c_str(
+                fmt::format("<a href='standardese://{}/'>{}</a>", unique_name.c_str(), str.c_str())
+                    .c_str());
             return *this;
         }
 
         code_block_writer& operator<<(char c)
         {
-            stream_.write_char(c);
+            write_str_escaped(&c, 1);
             return *this;
         }
 
@@ -115,8 +135,32 @@ namespace standardese
         }
 
     private:
+        void write_c_str(const char* str)
+        {
+            stream_.write_str(str, std::strlen(str));
+        }
+
+        void write_str_escaped(const char* str, std::size_t size)
+        {
+            if (!use_advanced_)
+                stream_.write_str(str, size);
+            else
+                for (auto ptr = str; ptr != str + size; ++ptr)
+                {
+                    if (*ptr == '&')
+                        write_c_str("&amp;");
+                    else if (*ptr == '<')
+                        write_c_str("&lt;");
+                    else if (*ptr == '>')
+                        write_c_str("&gt;");
+                    else
+                        stream_.write_char(*ptr);
+                }
+        }
+
         string_output    stream_;
         const md_entity* parent_;
+        bool             use_advanced_;
     };
 
     class md_document;