+++ src/template_processor.cpp
@@ -507,7 +507,8 @@ namespace
             return nullptr;
 
         auto              doc = md_document::make("");
-        code_block_writer writer(*doc);
+        code_block_writer writer(*doc, vars.get_parser().get_output_config().is_set(
+                                           output_flag::use_advanced_code_block));
         entity->generate_synopsis(vars.get_parser(), writer);
         doc->add_entity(writer.get_code_block());
 
@@ -657,7 +658,7 @@ raw_document standardese::process_template(const parser& p, const index& i,
         {
             auto entity = s.lookup_var(read_arg(ptr, last));
             if (entity)
-                s.get_buffer() += entity->get_index_name(true).c_str();
+                s.get_buffer() += entity->get_index_name(true, true).c_str();
             break;
         }
         case template_command::module: