+++ tool/main.cpp
@@ -254,14 +254,18 @@ int main(int argc, char* argv[])
              "the tab width (i.e. number of spaces, won't emit tab) of the code in the synthesis")
             ("output.width", po::value<unsigned>()->default_value(terminal_width),
              "the width of the output (used in e.g. commonmark format)")
-            ("output.show_complex_noexcept", po::value<bool>()->default_value(true)->implicit_value(true),
-            "whether or not complex noexcept expressions will be shown in the synopsis or replaced by \"see below\"")
             ("output.inline_doc", po::value<bool>()->default_value(true)->implicit_value(true),
              "whether or not some entity documentation (parameters etc.) will be shown inline")
+            ("output.advanced_code_block", po::value<bool>()->default_value(true)->implicit_value(true),
+            "whether or not an advanced (HTML) code block will be used")
+            ("output.show_complex_noexcept", po::value<bool>()->default_value(true)->implicit_value(true),
+             "whether or not complex noexcept expressions will be shown in the synopsis or replaced by \"see below\"")
             ("output.show_macro_replacement", po::value<bool>()->default_value(false)->implicit_value(true),
              "whether or not the replacement of macros will be shown")
             ("output.show_group_member_id", po::value<bool>()->default_value(true)->implicit_value(true),
              "whether or not to show the index of member group members in the synopsis")
+            ("output.show_group_output_section", po::value<bool>()->default_value(true)->implicit_value(true),
+            "whether or not member groups have an implicit output section")
             ("output.show_modules", po::value<bool>()->default_value(true)->implicit_value(true),
             "whether or not the module of an entity is shown in the documentation");
     // clang-format on